--- src/components/TempComponent/style.jsx
@@ -1,8 +1,0 @@
-import styled from "styled-components";
-
-export const TempContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-`;
