+++ src/components/index.js
@@ -1,8 +1,5 @@
 // 여기서 컴포넌트를 한번에 모아서 불러오는 것을 추천합니다! (절대 경로를 설정했기때문에 componetns만 해도 불러올수 있어요!)
 
-// Temp
-export { TempComponent } from "./TempComponent";
-
 /**
  * atoms
  */
@@ -18,7 +15,9 @@ export { ReceiptPaperInfo } from "./atoms/ReceiptPaperInfo";
 export { ReceiptBarcode } from "./atoms/ReceiptBarcode";
 export { ReceiptQuotes } from "./atoms/ReceiptQuotes";
 
+// common
 export { BackBtn } from "./atoms/BackBtn";
+
 /**
  * molecules
  */
@@ -50,7 +49,9 @@ export { MySection } from "./organisms/MySection";
 export { LoadingModal } from "./organisms/LoadingModal";
 export { AlertModal } from "./organisms/AlertModal";
 
-// templates
+/**
+ * templates
+ */
 export { ModalTemplate } from "./templates/ModalTemplate";
 
 /**
