+++ src/components/pages/AuthPage/AuthPage.jsx
@@ -3,7 +3,7 @@ import { useNavigate } from "react-router-dom";
 import Loading from "react-loading";
 import AtuhContext from "store/auth-context";
 import { ReactComponent as KakaoIcon } from "assets/kakao/KakaoTalk_logo.svg";
-import * as S from "./style";
+import * as S from "./AuthPage.styles";
 import { getKakaoToken } from "controllers/userController";
 
 export function AuthPage() {
