+++ src/components/pages/AuthPage/index.js
@@ -1,0 +1,1 @@
+export { AuthPage } from "./AuthPage";
