+++ src/components/pages/LoginPage/index.js
@@ -1,0 +1,1 @@
+export { LoginPage } from "./LoginPage";
