+++ src/components/pages/ReceiptPage/ReceiptPage.jsx
@@ -1,6 +1,5 @@
 import { useEffect, useRef, useState, useContext } from "react";
 import { useNavigate, useLocation } from "react-router-dom";
-import styled from "styled-components";
 import html2canvas from "html2canvas";
 import { getUserInfo } from "controllers/userController";
 import AtuhContext from "store/auth-context";
@@ -9,6 +8,7 @@ import { ReactComponent as SaveIcon } from "assets/receiptPage/save_icon.svg";
 import { ReactComponent as ShareIcon } from "assets/receiptPage/share_icon.svg";
 import { ReactComponent as BackIcon } from "assets/receiptPage/back_icon.svg";
 import dayjs from "dayjs";
+import * as S from "./ReceiptPage.styles";
 
 export function ReceiptPage() {
   const {
@@ -59,14 +59,14 @@ export function ReceiptPage() {
   }, []);
 
   return (
-    <Container>
-      <BackIconContainer onClick={() => navigate(-1)}>
+    <S.Container>
+      <S.BackIconContainer onClick={() => navigate(-1)}>
         <BackIcon />
-      </BackIconContainer>
-      <ReceiptContainer ref={receiptRef} scale={scale}>
+      </S.BackIconContainer>
+      <S.ReceiptContainer ref={receiptRef} scale={scale}>
         <ReceiptPaper todos={todos} />
-      </ReceiptContainer>
-      <IconContainer>
+      </S.ReceiptContainer>
+      <S.IconContainer>
         <div onClick={handleShare}>
           <ShareIcon />
           <span>SHARE</span>
@@ -75,75 +75,8 @@ export function ReceiptPage() {
           <SaveIcon />
           <span>SAVE</span>
         </div>
-      </IconContainer>
+      </S.IconContainer>
       {modalOn && <AlertModal onClick={() => setModalOn(false)} />}
-    </Container>
+    </S.Container>
   );
 }
-
-const Container = styled.div`
-  width: 100%;
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  background-color: ${(props) => props.theme.bk};
-`;
-
-const BackIconContainer = styled.div`
-  width: 36px;
-  margin-top: 46px;
-  margin-left: 14px;
-  padding: 8px;
-  opacity: 50%;
-  cursor: pointer;
-
-  @media screen and (min-width: 768px) {
-    position: absolute;
-    top: 42px;
-    left: 48px;
-  }
-`;
-
-const ReceiptContainer = styled.div`
-  width: 100%;
-  height: 70%;
-  margin-top: 5%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-
-  > div {
-    transform: scale(${(props) => props.scale});
-  }
-`;
-
-const IconContainer = styled.div`
-  width: 100%;
-  height: 12%;
-  display: flex;
-  align-self: center;
-  align-items: center;
-  justify-content: center;
-  position: absolute;
-  bottom: 24px;
-
-  div {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    padding: 4px;
-    margin: 4px 24px;
-    cursor: pointer;
-  }
-
-  svg {
-    padding: 8px;
-    opacity: 40%;
-  }
-
-  span {
-    font-family: "Courier Prime", monospace;
-    font-size: 16px;
-    color: #efefef;
-  }
-`;
