+++ src/components/pages/ReceiptPage/index.js
@@ -1,0 +1,1 @@
+export { ReceiptPage } from "./ReceiptPage";
