+++ src/components/pages/TodoPage/index.js
@@ -1,0 +1,1 @@
+export { TodoPage } from "./TodoPage";
