+++ src/components/pages/UserPage/index.js
@@ -1,0 +1,1 @@
+export { UserPage } from "./UserPage";
