+++ src/components/templates/ModalTemplate/index.js
@@ -1,0 +1,1 @@
+export { ModalTemplate } from "./ModalTemplate";
